+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # App Center SDK for Android Change Log
 
-## Version 1.11.0 (Under active development)
+## Version 1.10.1 (Under active development)
+
+### AppCenter
+
+* **[Fix]** Do not delete old logs when trying to add a log larger than the maximum storage capacity.
+
+### AppCenterCrashes
+
+* **[Fix]** Fix a bug where crash data file could leak when the database is full.
 
 ### AppCenterPush